requirements.txt
streamlit>=1.23.0
requests>=2.25.0
WooCommerce>=3.0.0
Pillow>=8.0.0
pandas>=1.3.0
cryptography>=3.0.0
lxml>=4.6.0
html5lib>=1.1
python-dateutil>=2.8.0
urllib3>=1.26.0
typing-extensions>=4.0.0